requirements.txt
evaluate
fasttext_langdetect==1.0.5
fitz==0.0.1.dev2
inscriptis==2.3.2
jsonlines==4.0.0
matplotlib
mitreattack_python==2.0.14
networkx==3.1
numpy
pandas
psutil==5.9.5
pyinflect==0.5.1
python_Levenshtein==0.25.1
regex
Requests
scikit_learn
snorkel==0.9.9
spacy==3.5.0
tensorflow_hub==0.16.1
tqdm==4.65.0
spacy-transformers==1.2.5
coreferee
